Fix: history nagivgation for protected routes.
- If users try to access a protected page, e.g., add project, they
will get redirected to login page, if they click on the go back button
in the browser, it won't work because they will get redirect once again
to the login page.
- Using `replace` instead of `push` will replace protected page from
browser history with the login page, so navigating back will be login
page -> the page the user tried accessing the protected page from.